pyproject.toml
[tool.poetry]
name = "card-index"
version = "0.1.0"
description = "Displays the employee hierarchy in tree form."
authors = ["Woolf <woolf.no1@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
litestar = {extras = ["jinja", "jwt"], version = "^2.12.1"}
tortoise-orm = {extras = ["asyncpg"], version = "^0.21.7"}
pydantic-settings = "^2.6.0"
pip = "^24.3.1"
uvicorn = "^0.32.0"
aerich = "^0.7.2"
faker = "^30.8.2"
passlib = "^1.7.4"
bcrypt = "^4.2.0"



[tool.aerich]
tortoise_orm = "db.db.TORTOISE_ORM"
location = "./migrations"
src_folder = "./."
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
